// src/types/draft-class.ts
export enum PlayerArchetype {
  Invalid_ = 0,
  QB_FieldGeneral = 1,
  QB_StrongArm = 2,
  QB_Scrambler = 3,
  HB_PowerBack = 4,
  HB_SpeedBack = 5,
  HB_ReceivingBack = 6,
  HB_BlockingBack = 7,
  FB_Blocking = 8,
  FB_Receiving = 9,
  WR_DeepThreat = 10,
  WR_Possession = 11,
  WR_Slot = 12,
  WR_Blocking = 13,
  TE_Receiving = 14,
  TE_Blocking = 15,
  TE_Balanced = 16,
  LT_PassBlocking = 17,
  LT_RunBlocking = 18,
  LG_PassBlocking = 19,
  LG_RunBlocking = 20,
  C_Balanced = 21,
  RG_PassBlocking = 22,
  RG_RunBlocking = 23,
  RT_PassBlocking = 24,
  RT_RunBlocking = 25,
  LE_PassRush = 26,
  LE_RunStuffing = 27,
  RE_PassRush = 28,
  RE_RunStuffing = 29,
  DT_PassRush = 30,
  DT_RunStuffing = 31,
  LOLB_PassRush = 32,
  LOLB_RunStuffing = 33,
  MLB_PassRush = 34,
  MLB_RunStuffing = 35,
  ROLB_PassRush = 36,
  ROLB_RunStuffing = 37,
  CB_ManToMan = 38,
  CB_ZoneCoverage = 39,
  CB_PressCoverage = 40,
  FS_DeepZone = 41,
  FS_ManToMan = 42,
  SS_RunSupport = 43,
  SS_ManToMan = 44,
  K_Accurate = 45,
  K_Power = 46,
  P_Power = 47,
  P_Accurate = 48,
}

export interface Prospect {
  // Personal Info
  id: string;
  firstName: string;
  lastName: string;
  homeState: number;
  homeTown: string;
  college: number;
  birthDate: number;
  age: number;
  heightInches: number;
  weight: number;
  position: number;
  archetype: PlayerArchetype;
  jerseyNum: number;
  assetName: string;

  // Draft Info
  draftable: number;
  draftPick: number;
  draftRound: number;

  // Ratings
  overall: number;
  acceleration: number;
  agility: number;
  awareness: number;
  ballCarrierVision: number;
  blockShedding: number;
  breakSack: number;
  breakTackle: number;
  carrying: number;
  catching: number;
  catchInTraffic: number;
  changeOfDirection: number;
  finesseMoves: number;
  hitPower: number;
  impactBlocking: number;
  injury: number;
  jukeMove: number;
  jumping: number;
  kickAccuracy: number;
  kickPower: number;
  kickReturn: number;
  leadBlock: number;
  manCoverage: number;
  passBlockFinesse: number;
  passBlockPower: number;
  passBlock: number;
  personality: number;
  playAction: number;
  playRecognition: number;
  powerMoves: number;
  pressCoverage: number;
  pursuit: number;
  release: number;
  shortRouteRunning: number;
  mediumRouteRunning: number;
  deepRouteRunning: number;
  runBlockFinesse: number;
  runBlockPower: number;
  runBlock: number;
  runningStyle: number;
  spectacularCatch: number;
  speed: number;
  spinMove: number;
  stamina: number;
  stiffArm: number;
  strength: number;
  tackle: number;
  throwAccuracyDeep: number;
  throwAccuracyMid: number;
  throwAccuracy: number;
  throwAccuracyShort: number;
  throwOnTheRun: number;
  throwPower: number;
  throwUnderPressure: number;
  toughness: number;
  trucking: number;
  zoneCoverage: number;
  morale: number;

  // Traits
  traitBigHitter: number;
  traitPossessionCatch: number;
  traitClutch: number;
  traitCoverBall: number;
  traitDeepBall: number;
  traitDlBullRush: number;
  traitDlSpinMove: number;
  traitDlSwimMove: number;
  traitDropsOpen: number;
  traitSidelineCatch: number;
  traitFightForYards: number;
  traitUnk1: number;
  traitHighMotor: number;
  traitAggressiveCatch: number;
  traitPenalty: number;
  traitPlayBall: number;
  traitPumpFake: number;
  traitLbStyle: number;
  traitSensePressure: number;
  traitUnk2: number;
  traitStripBall: number;
  traitTackleLow: number;
  traitThrowAway: number;
  traitTightSpiral: number;
  traitTendency: number;
  traitRunAfterCatch: number;
  devTrait: number;
  traitPredictability: number;

  // Appearance
  visuals: string;
  genericHead: GenericHead;
  handedness: number;
  portraitId: number;
  qbStyle: number;
  qbStance: number;
  unk3: number;
  unk4: number;
  unk5: number;
  unk6: number;
  visMoveType: number;
  unk8: number;
  commentaryId: number;
  unkByte2: number;
}